+++ src/Routes.jsx
@@ -1,10 +1,8 @@
 import React from "react";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
-import Membres from "./pages/Membres";
+import {BrowserRouter, Switch, Route} from "react-router-dom";
 import HomepageLayout from "./pages/HomeLayout";
 import Connexion from "./pages/Connexion";
 import Register from "./pages/Register";
-import MyProfile from "./pages/MyProfile"
 
 /**
  * TODO : Make each component
@@ -12,18 +10,13 @@ import MyProfile from "./pages/MyProfile"
  */
 
 export default () => {
-  return (
-    <BrowserRouter>
-      <Switch>
-        <Route exact path="/" component={HomepageLayout} />
-        <Route exact path="/membres" component={Membres} />
-        <Route exact path="/connexion" component={Connexion} />
-        <Route exact path="/enregistrement" component={Register} />
-          <Route exact path="/myprofile" component={MyProfile}
-        {/* <Route exact path="/annuaire" component={Annuaire} />
-        <Route exact path="/entreprise" component={Entreprise} />
-        <Route exact path="/enregistrement" component={Register} /> */}
-      </Switch>
-    </BrowserRouter>
-  );
+    return (
+        <BrowserRouter>
+            <Switch>
+                <Route exact path="/" component={HomepageLayout}/>
+                <Route exact path="/connexion" component={Connexion}/>
+                <Route exact path="/enregistrement" component={Register}/>
+            </Switch>
+        </BrowserRouter>
+    );
 };
