+++ src/components/Homepage/HomeContent.jsx
@@ -27,7 +27,7 @@ const HomeContent = () => (
                         needs... through pure data analytics.
                     </p>
                     <Header as="h3" style={{fontSize: "2em"}}>
-                        We Make Bananas That Can Dance
+                        Content
                     </Header>
                     <p style={{fontSize: "1.33em"}}>
                         Yes that's right, you thought it was the stuff of dreams, but even
