+++ src/components/NavBar/ConnectedNavDesktop.jsx
@@ -12,6 +12,7 @@ import {
 
 import "./NavBar.css";
 import logoLinear from "./logo_linear.jpeg";
+import * as Auth from "../../api/Auth";
 
 /* Heads up!
  * Neither Semantic UI nor Semantic UI React offer a responsive navbar, however, it can be implemented easily.
@@ -20,8 +21,19 @@ import logoLinear from "./logo_linear.jpeg";
 export default class DesktopConnectedNavBarContainer extends Component {
     state = {};
 
+    constructor(props) {
+        super(props);
+        this.hideFixedMenu = this.hideFixedMenu.bind(this);
+        this.showFixedMenu = this.showFixedMenu.bind(this);
+        this.logOut = this.logOut.bind(this);
+    }
+
     hideFixedMenu = () => this.setState({fixed: false});
     showFixedMenu = () => this.setState({fixed: true});
+
+    logOut = () => {
+        Auth.removeToken();
+    };
 
     render() {
         const {children} = this.props;
@@ -45,32 +57,31 @@ export default class DesktopConnectedNavBarContainer extends Component {
                             inverted={!fixed}
                             pointing={!fixed}
                             secondary={!fixed}
-                            size="medium"
+                            size="large"
                         >
                             <Container fluid>
                                 <Image size='tiny' src={logoLinear} style={{ marginRight: '1.5em' }} wrapped/>
                                 <Menu.Item as="a" active href="./" color="red">
                                     Accueil
                                 </Menu.Item>
-                                <Menu.Item as="a" active href="./enregistrement" color="green">
+                                <Menu.Item as="a" active href="./" color="green">
                                     Entreprises
                                 </Menu.Item>
-                                <Menu.Item as="a" active href="./membres" color="red">
+                                <Menu.Item as="a" active href="./" color="red">
                                     Membres
                                 </Menu.Item>
-                                <Menu.Item as="a" active href="./annuaire" color="green">
+                                <Menu.Item as="a" active href="./" color="green">
                                     Annuaire
                                 </Menu.Item>
                                 <Menu.Item position="right">
-                                    <Button as="a" inverted={!fixed} href="./connexion">
-                                        Connectez-vous
+                                    <Button as="a" inverted={!fixed} href="./myprofile">
+                                        Mon Profil
                                     </Button>
                                     <Button
-                                        as="a"
-                                        inverted={!fixed}
-                                        style={{marginLeft: "0.5em"}}
-                                    >
-                                        Enregistrez-vous
+                                        onClick={this.logOut}
+                                        as="a" inverted={!fixed}
+                                        style={{marginLeft: "0.5em"}}>
+                                        Se Déconnecter
                                     </Button>
                                 </Menu.Item>
                             </Container>
