+++ src/pages/MyProfile.jsx
@@ -1,0 +1,8 @@
+import React, {Component} from "react";
+
+export default class MyProfile extends Component {
+
+    render() {
+        return (<h1>My Profile</h1>);
+    }
+}
