+++ src/pages/Register.jsx
@@ -1,48 +1,64 @@
-import React from "react";
-import { Form, Input, TextArea, Button, Select } from "semantic-ui-react";
+import React, {Component} from "react";
+import {Form, Input, TextArea, Button, Select, Radio} from "semantic-ui-react";
 
-const genderOptions = [
-  { key: "m", text: "Male", value: "male" },
-  { key: "f", text: "Female", value: "female" }
-];
+export default class Register extends Component {
 
-const FormExampleFieldControlId = () => (
-  <Form>
-    <Form.Group widths="equal">
-      <Form.Field
-        id="form-input-control-first-name"
-        control={Input}
-        label="First name"
-        placeholder="First name"
-      />
-      <Form.Field
-        id="form-input-control-last-name"
-        control={Input}
-        label="Last name"
-        placeholder="Last name"
-      />
-      <Form.Field
-        control={Select}
-        options={genderOptions}
-        label={{ children: "Gender", htmlFor: "form-select-control-gender" }}
-        placeholder="Gender"
-        search
-        searchInput={{ id: "form-select-control-gender" }}
-      />
-    </Form.Group>
-    <Form.Field
-      id="form-textarea-control-opinion"
-      control={TextArea}
-      label="Opinion"
-      placeholder="Opinion"
-    />
-    <Form.Field
-      id="form-button-control-public"
-      control={Button}
-      content="Confirm"
-      label="Label with htmlFor"
-    />
-  </Form>
-);
+    render() {
+        return(
 
-export default FormExampleFieldControlId;
+            <Form>
+                <Form.Group widths="equal">
+                    <Form.Field>
+                        <Radio
+                            label='Membre'
+                            name='radioGroup'
+                            value='member'
+                            checked={this.state.value === 'member'}
+                            onChange={this.handleRadioChange}
+                        />
+                    </Form.Field>
+                    <Form.Field>
+                        <Radio
+                            label='Entreprise'
+                            name='radioGroup'
+                            value='company'
+                            checked={this.state.value === 'company'}
+                            onChange={this.handleRadioChange}
+                        />
+                    </Form.Field>
+                    <Form.Field
+                        id="form-input-control-first-name"
+                        control={Input}
+                        label="First name"
+                        placeholder="First name"
+                    />
+                    <Form.Field
+                        id="form-input-control-last-name"
+                        control={Input}
+                        label="Last name"
+                        placeholder="Last name"
+                    />
+                    <Form.Field
+                        id="form-input-control-last-name"
+                        control={Input}
+                        label="Last name"
+                        placeholder="Last name"
+                    />
+                    <Form.Field
+                        id="form-input-control-last-name"
+                        control={Input}
+                        label="Last name"
+                        placeholder="Last name"
+                    />
+                </Form.Group>
+                <Form.Field
+                    id="form-button-control-public"
+                    control={Button}
+                    content="Confirm"
+                    label="Label with htmlFor"
+                />
+            </Form>
+
+        );
+    }
+}
